Eliminate style warnings.
There were style warnings about incorrect inlining for random (in
protocol.lisp) and for seed (in generator.lisp).

The former happens because the functions invoked by random (which is
declared inline) are not defined before random.  SBCL does not like
this.  Fixed by moving RANDOM to the bottom of the file.

The latter happens because a compiler-macro is declared for seed after
a call is compiled. But ... there does not seem to be any way for me
to tell where that call is (or calls are). So I simply quash this
style-warning.